+++ src/ingame/entry/ap_bgost.h
@@ -3,13 +3,15 @@
 
 #include "typedefs.h"
 
+#include <stdint.h>
+
 typedef struct {
     u_char load_mode;
     u_char rel_mot;
     u_char bg_no;
-    u_int mdl_adr;
-    u_int mot_adr;
-    u_int se_adr;
+    uint64_t mdl_adr;
+    uint64_t mot_adr;
+    uint64_t se_adr;
 } BGST_WRK;
 
 // extern BGST_WRK bgst_wrk;
