+++ src/ingame/map/map_area.c
@@ -1,6 +1,8 @@
 #include "common.h"
 #include "map_area.h"
 
+#include <stdint.h>
+
 /* data 332120 */ u_char area_dat[11][6];
 /* sdata 3570ec */ u_char one_room_load_ok;
 /* data 332170 */ AREA_WRK area_wrk;
@@ -37,6 +39,6 @@ int AreaRoomAllLoad(u_char area_no)
 {
 }
 
-int GetEmptyRoomAddr()
+int64_t GetEmptyRoomAddr()
 {
 }
